Moved DeconzEvent to sensor since it is only Switch buttonpress that will be sent as event
